==> R/functionalPlantIndicators/code naturally open/6 post scaling analyses_natopen_GRUK reftest.R <==
#### getting all the data from previous scripts in ####
load("P:/41201785_okologisk_tilstand_2022_2023/data/FPI_output large files for markdown/data_natopen.RData")
load("P:/41201785_okologisk_tilstand_2022_2023/data/FPI_output large files for markdown/ref_lists_natopen.RData")
load("P:/41201785_okologisk_tilstand_2022_2023/data/FPI_output large files for markdown/results.natopen.GRUK.reftest.RData")


#### comparing the references ####
par(mfrow=c(2,6))
plot(density(natopen.ref.cov[["CC"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="Reference comparison T2-C7",xlab="CSR-C")
points(density(natopen.ref.cov[["CC"]]$"T2-C7_BN"),type="l",lty=2)
legend("topright", legend=c("regular","BN"), lty=c(1,2))

plot(density(natopen.ref.cov[["SS"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="",xlab="CSR-S")
points(density(natopen.ref.cov[["SS"]]$"T2-C7_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["RR"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="",xlab="CSR-R")
points(density(natopen.ref.cov[["RR"]]$"T2-C7_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["Light"]]$"T2-C7"),type="l",lty=1,xlim=c(0,7),main="",xlab="Light")
points(density(natopen.ref.cov[["Light"]]$"T2-C7_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["Nitrogen"]]$"T2-C7"),type="l",lty=1,xlim=c(0,9),main="",xlab="Nitrogen")
points(density(natopen.ref.cov[["Nitrogen"]]$"T2-C7_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C7"),type="l",lty=1,xlim=c(0,9),main="",xlab="Soil_disturbance")
points(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C7_BN"),type="l",lty=2)


plot(density(natopen.ref.cov[["CC"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="Reference comparison T2-C8",xlab="CSR-C")
points(density(natopen.ref.cov[["CC"]]$"T2-C8_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["SS"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="",xlab="CSR-S")
points(density(natopen.ref.cov[["SS"]]$"T2-C8_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["RR"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="",xlab="CSR-R")
points(density(natopen.ref.cov[["RR"]]$"T2-C8_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["Light"]]$"T2-C8"),type="l",lty=1,xlim=c(0,7),main="",xlab="Light")
points(density(natopen.ref.cov[["Light"]]$"T2-C8_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["Nitrogen"]]$"T2-C8"),type="l",lty=1,xlim=c(0,9),main="",xlab="Nitrogen")
points(density(natopen.ref.cov[["Nitrogen"]]$"T2-C8_BN"),type="l",lty=2)
plot(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C8"),type="l",lty=1,xlim=c(0,9),main="",xlab="Soil_disturbance")
points(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C8_BN"),type="l",lty=2)

#### references vs. original values ####
x11()
par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["CC"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="T2-C7 data vs. reference",xlab="CSR-C")
points(density(natopen.ref.cov[["CC"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","CC1"], na.rm = TRUE),type="l",lty=1,col="red")
plot(density(natopen.ref.cov[["CC"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="T2-C8 data vs. reference",xlab="CSR-C")
points(density(natopen.ref.cov[["CC"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","CC1"], na.rm = TRUE),type="l",lty=1,col="red")
legend("topright", legend=c("regular","BN","GRUK data"), lty=c(1,2,1), col=c("black","black","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["SS"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="T2-C7 data vs. reference",xlab="CSR-S")
points(density(natopen.ref.cov[["SS"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","SS1"], na.rm = TRUE),type="l",lty=1,col="red")
plot(density(natopen.ref.cov[["SS"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="T2-C8 data vs. reference",xlab="CSR-S")
points(density(natopen.ref.cov[["SS"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","SS1"], na.rm = TRUE),type="l",lty=1,col="red")
legend("topright", legend=c("regular","BN","GRUK data"), lty=c(1,2,1), col=c("black","black","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["RR"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="T2-C7 data vs. reference",xlab="CSR-R")
points(density(natopen.ref.cov[["RR"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","RR1"], na.rm = TRUE),type="l",lty=1,col="red")
plot(density(natopen.ref.cov[["RR"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="T2-C8 data vs. reference",xlab="CSR-R")
points(density(natopen.ref.cov[["RR"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","RR1"], na.rm = TRUE),type="l",lty=1,col="red")
legend("topright", legend=c("regular","BN","GRUK data"), lty=c(1,2,1), col=c("black","black","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["Light"]]$"T2-C7"),type="l",lty=1,xlim=c(0,7),main="T2-C7 data vs. reference",xlab="Light")
points(density(natopen.ref.cov[["Light"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","Light1"], na.rm = TRUE),type="l",lty=1,col="red")
plot(density(natopen.ref.cov[["Light"]]$"T2-C8"),type="l",lty=1,xlim=c(0,7),main="T2-C8 data vs. reference",xlab="Light")
points(density(natopen.ref.cov[["Light"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","Light1"], na.rm = TRUE),type="l",lty=1,col="red")
legend("topleft", legend=c("regular","BN","GRUK data"), lty=c(1,2,1), col=c("black","black","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["Nitrogen"]]$"T2-C7"),type="l",lty=1,xlim=c(0,9),main="T2-C7 data vs. reference",xlab="Nitrogen")
points(density(natopen.ref.cov[["Nitrogen"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","Nitrogen1"], na.rm = TRUE),type="l",lty=1,col="red")
plot(density(natopen.ref.cov[["Nitrogen"]]$"T2-C8"),type="l",lty=1,xlim=c(0,9),main="T2-C8 data vs. reference",xlab="Nitrogen")
points(density(natopen.ref.cov[["Nitrogen"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","Nitrogen1"], na.rm = TRUE),type="l",lty=1,col="red")
legend("topright", legend=c("regular","BN","GRUK data"), lty=c(1,2,1), col=c("black","black","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C7"),type="l",lty=1,xlim=c(0,9),main="T2-C7 data vs. reference",xlab="Soil_disturbance")
points(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","Soil_disturbance1"], na.rm = TRUE),type="l",lty=1,col="red")
plot(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C8"),type="l",lty=1,xlim=c(0,9),main="T2-C8 data vs. reference",xlab="Soil_disturbance")
points(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular","Soil_disturbance1"], na.rm = TRUE),type="l",lty=1,col="red")
legend("topright", legend=c("regular","BN","GRUK data"), lty=c(1,2,1), col=c("black","black","red"))


#### condition estimates ####
results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering <- as.factor(results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering)

x11()
par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["CC"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="T2-C7 data vs. reference",xlab="CSR-C")
points(density(natopen.ref.cov[["CC"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","CC1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","CC1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","CC1"], na.rm = TRUE),type="l",lty=3,col="red")

plot(density(natopen.ref.cov[["CC"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="T2-C8 data vs. reference",xlab="CSR-C")
points(density(natopen.ref.cov[["CC"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","CC1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","CC1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","CC1"], na.rm = TRUE),type="l",lty=3,col="red")
legend("topright", legend=c("ref regular","ref BN","GRUK tilst=0", "GRUK tilst=1", "GRUK tilst=2"), lty=c(1,2,1,2,3), col=c("black","black","red","red","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["SS"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="T2-C7 data vs. reference",xlab="CSR-S")
points(density(natopen.ref.cov[["SS"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","SS1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","SS1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","SS1"], na.rm = TRUE),type="l",lty=3,col="red")

plot(density(natopen.ref.cov[["SS"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="T2-C8 data vs. reference",xlab="CSR-S")
points(density(natopen.ref.cov[["SS"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","SS1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","SS1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","SS1"], na.rm = TRUE),type="l",lty=3,col="red")
legend("topright", legend=c("ref regular","ref BN","GRUK tilst=0", "GRUK tilst=1", "GRUK tilst=2"), lty=c(1,2,1,2,3), col=c("black","black","red","red","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["RR"]]$"T2-C7"),type="l",lty=1,xlim=c(0,1),main="T2-C7 data vs. reference",xlab="CSR-R")
points(density(natopen.ref.cov[["RR"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","RR1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","RR1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","RR1"], na.rm = TRUE),type="l",lty=3,col="red")

plot(density(natopen.ref.cov[["RR"]]$"T2-C8"),type="l",lty=1,xlim=c(0,1),main="T2-C8 data vs. reference",xlab="CSR-R")
points(density(natopen.ref.cov[["RR"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","RR1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","RR1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","RR1"], na.rm = TRUE),type="l",lty=3,col="red")
legend("topright", legend=c("ref regular","ref BN","GRUK tilst=0", "GRUK tilst=1", "GRUK tilst=2"), lty=c(1,2,1,2,3), col=c("black","black","red","red","red"))

par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["Light"]]$"T2-C7"),type="l",lty=1,xlim=c(0,7),main="T2-C7 data vs. reference",xlab="Light")
points(density(natopen.ref.cov[["Light"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","Light1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","Light1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","Light1"], na.rm = TRUE),type="l",lty=3,col="red")

plot(density(natopen.ref.cov[["Light"]]$"T2-C8"),type="l",lty=1,xlim=c(0,7),main="T2-C8 data vs. reference",xlab="Light")
points(density(natopen.ref.cov[["Light"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","Light1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","Light1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","Light1"], na.rm = TRUE),type="l",lty=3,col="red")
legend("topleft", legend=c("ref regular","ref BN","GRUK tilst=0", "GRUK tilst=1", "GRUK tilst=2"), lty=c(1,2,1,2,3), col=c("black","black","red","red","red"))


par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["Nitrogen"]]$"T2-C7"),type="l",lty=1,xlim=c(0,9),main="T2-C7 data vs. reference",xlab="Nitrogen")
points(density(natopen.ref.cov[["Nitrogen"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","Nitrogen1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","Nitrogen1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","Nitrogen1"], na.rm = TRUE),type="l",lty=3,col="red")

plot(density(natopen.ref.cov[["Nitrogen"]]$"T2-C8"),type="l",lty=1,xlim=c(0,9),main="T2-C8 data vs. reference",xlab="Nitrogen")
points(density(natopen.ref.cov[["Nitrogen"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","Nitrogen1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","Nitrogen1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","Nitrogen1"], na.rm = TRUE),type="l",lty=3,col="red")
legend("topright", legend=c("ref regular","ref BN","GRUK tilst=0", "GRUK tilst=1", "GRUK tilst=2"), lty=c(1,2,1,2,3), col=c("black","black","red","red","red"))


par(mfrow=c(1,2))
plot(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C7"),type="l",lty=1,xlim=c(0,9),main="T2-C7 data vs. reference",xlab="Soil_disturbance")
points(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C7_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","Soil_disturbance1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","Soil_disturbance1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-7" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","Soil_disturbance1"], na.rm = TRUE),type="l",lty=3,col="red")

plot(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C8"),type="l",lty=1,xlim=c(0,9),main="T2-C8 data vs. reference",xlab="Soil_disturbance")
points(density(natopen.ref.cov[["Soil_disturbance"]]$"T2-C8_BN"),type="l",lty=2)
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="0","Soil_disturbance1"], na.rm = TRUE),type="l",lty=1,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="1","Soil_disturbance1"], na.rm = TRUE),type="l",lty=2,col="red")
points(density(results.natopen.GRUK.reftest[["original"]][results.natopen.GRUK.reftest[["original"]]$Kartleggingsenhet=="T2-C-8" & results.natopen.GRUK.reftest[["original"]]$reference_type=="regular" & results.natopen.GRUK.reftest[["original"]]$tilstandsvurdering=="2","Soil_disturbance1"], na.rm = TRUE),type="l",lty=3,col="red")
legend("topleft", legend=c("ref regular","ref BN","GRUK tilst=0", "GRUK tilst=1", "GRUK tilst=2"), lty=c(1,2,1,2,3), col=c("black","black","red","red","red"))
